+++ extmod/nimble/modbluetooth_nimble.h
@@ -24,8 +24,8 @@
  * THE SOFTWARE.
  */
 
-#ifndef MICROPY_INCLUDED_EXTMOD_MODBLUETOOTH_NIMBLE_H
-#define MICROPY_INCLUDED_EXTMOD_MODBLUETOOTH_NIMBLE_H
+#ifndef MICROPY_INCLUDED_EXTMOD_NIMBLE_MODBLUETOOTH_NIMBLE_H
+#define MICROPY_INCLUDED_EXTMOD_NIMBLE_MODBLUETOOTH_NIMBLE_H
 
 #define MP_BLUETOOTH_NIMBLE_MAX_SERVICES (8)
 
@@ -51,4 +51,4 @@ void mp_bluetooth_nimble_port_postinit(void);
 void mp_bluetooth_nimble_port_deinit(void);
 void mp_bluetooth_nimble_port_start(void);
 
-#endif // MICROPY_INCLUDED_EXTMOD_MODBLUETOOTH_NIMBLE_H
+#endif // MICROPY_INCLUDED_EXTMOD_NIMBLE_MODBLUETOOTH_NIMBLE_H
