+++ extmod/nimble/nimble.mk
@@ -1,9 +1,20 @@
-# Makefile directives for Apache mynewt nimble BLE component
+# Makefile directives for Apache Mynewt NimBLE component
 
 ifeq ($(MICROPY_BLUETOOTH_NIMBLE),1)
 
-NIMBLE_LIB_DIR = lib/mynewt-nimble
+EXTMOD_SRC_C += extmod/nimble/modbluetooth_nimble.c
+
+CFLAGS_MOD += -DMICROPY_BLUETOOTH_NIMBLE=1
+
 NIMBLE_EXTMOD_DIR = extmod/nimble
+
+# Use NimBLE from the submodule in lib/mynewt-nimble by default,
+# allowing a port to use their own system version (e.g. ESP32).
+MICROPY_BLUETOOTH_NIMBLE_BINDINGS_ONLY ?= 0
+
+ifeq ($(MICROPY_BLUETOOTH_NIMBLE_BINDINGS_ONLY),0)
+
+NIMBLE_LIB_DIR = lib/mynewt-nimble
 
 SRC_LIB += $(addprefix $(NIMBLE_LIB_DIR)/, \
 	$(addprefix ext/tinycrypt/src/, \
@@ -76,8 +87,6 @@ EXTMOD_SRC_C += $(addprefix $(NIMBLE_EXTMOD_DIR)/, \
 	nimble/hci_uart.c \
 	)
 
-CFLAGS_MOD += -DMICROPY_BLUETOOTH_NIMBLE=1
-
 INC += -I$(TOP)/$(NIMBLE_EXTMOD_DIR)
 INC += -I$(TOP)/$(NIMBLE_LIB_DIR)
 INC += -I$(TOP)/$(NIMBLE_LIB_DIR)/ext/tinycrypt/include
@@ -91,5 +100,7 @@ INC += -I$(TOP)/$(NIMBLE_LIB_DIR)/nimble/transport/uart/include
 INC += -I$(TOP)/$(NIMBLE_LIB_DIR)/porting/nimble/include
 
 $(BUILD)/$(NIMBLE_LIB_DIR)/%.o: CFLAGS += -Wno-maybe-uninitialized -Wno-pointer-arith -Wno-unused-but-set-variable -Wno-format
+
+endif
 
 endif
